requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai>=1.0.0
huggingface_hub
aiohttp
botbuilder-core
botbuilder-schema
